+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 ### Changed
 
 - Updated the Turkish localization file (by @NovaDiablox)
+- Keyhelp and weapon HUD Help now use the global scale factor
 
 ### Fixed